requirements.txt
sentence-transformers
torch
transformers
scikit-learn
numpy
tqdm
qdrant-client
streamlit
PyPDF2
pdf2image
pytesseract
Pillow
python-dotenv
keybert